backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
nltk==3.8.1
beautifulsoup4
requests
sqlalchemy
psycopg2-binary
python-multipart
rapidfuzz
python-dotenv
